requirements.txt
Flask
git+https://github.com/openai/whisper.git
easyocr
Pillow
numpy
pdf2image